+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ember-cli-fastboot changelog
 
+## Unreleased
+
+### Added
+* allow reading sandboxGlobals from app's config/environment
+
 ## 1.1.0
 
 * Bumping `fastboot-express-middleware` to 1.1.0